miniProject

+++ src/miniProject/vo/ordersVO.java
@@ -1,17 +1,17 @@
 package miniProject.vo;
-
+import java.math.BigDecimal;
 import java.sql.Date;
 
 // 주문 정보 테이블
 public class ordersVO {
     private String order_id;
     private String customer_id;
-    private double total_cost;
+    private BigDecimal total_cost;
     private Date order_date;
     private String payment_method;
     private String shipping_address;
 
-    public ordersVO(String order_id, String customer_id, double total_cost, Date order_date,
+    public ordersVO(String order_id, String customer_id, BigDecimal total_cost, Date order_date,
                     String payment_method, String shipping_address) {
         this.order_id = order_id;
         this.customer_id = customer_id;
@@ -37,11 +37,11 @@ public class ordersVO {
         this.customer_id = customer_id;
     }
 
-    public double getTotal_cost() {
+    public BigDecimal getTotal_cost() {
         return total_cost;
     }
 
-    public void setTotal_cost(double total_cost) {
+    public void setTotal_cost(BigDecimal total_cost) {
         this.total_cost = total_cost;
     }
